Pull in patch from current: Errata: The ef(4) driver will complain when adding an address with ifconfig (ifconfig: SIOCAIFADDR: Invalid argument). Fix: add missing 'break'

+++ gnu/usr.bin/binutils/binutils/wrstabs.c
@@ -1099,8 +1099,7 @@ stab_array_type (p, low, high, stringp)
       sprintf (buf, "%ld=@S;", index);
     }
 
-  sprintf (buf + strlen (buf), "ar%s;%ld;%ld;%s",
-	   range, (long) low, (long) high, element);
+  sprintf (buf + strlen (buf), "ar%s;%ld;%ld;%s", range, low, high, element);
   free (range);
   free (element);
